Add client testing for emails
