+++ babylab/app/templates/apt.html
@@ -83,7 +83,7 @@
             <th>
               Is taxi booked?&nbsp;&nbsp;&nbsp;&nbsp;<i class="fa fa-check"></i>
             </th>
-            <td>{{ data["taxi_isbooked"] }}</td>
+            <td>{{ data["taxi_isbooked"] | safe }}</td>
           </tr>
           <tr>
             <th>
