+++ babylab/app/templates/apt_modify.html
@@ -12,8 +12,8 @@
 <header>
   <div class="jumbotron jumbotron-fluid">
     <div class="container">
-      <h1 class="display-4">Appointments ({{ appt_id }})</h1>
-      <p class="lead">Modifying details about appointment {{ appt_id }}. Hit the "Submit" button when finished.</p>
+      <h1 class="display-4">Appointments ({{ apt_id }})</h1>
+      <p class="lead">Modifying details about appointment {{ apt_id }}. Hit the "Submit" button when finished.</p>
     </div>
   </div>
   <hr class="my-4">
@@ -35,7 +35,7 @@
       <div class="card-body">
         <div class="col">
           <input type="hidden" class="form-group" name="inputId" value="{{ ppt_id }}" readonly>
-          <input type="hidden" class="form-group" name="inputAptId" value="{{ appt_id }}" readonly>
+          <input type="hidden" class="form-group" name="inputAptId" value="{{ apt_id }}" readonly>
           <label for="inputStudy" class="form-label">
             Study&nbsp;&nbsp;<i class="fa  fa-flask"></i>
           </label>
