+++ tests/functional/test_appointments.py
@@ -1,5 +1,12 @@
 """Test appointments endpoints."""
 
+import os
+import time
+import pytest
+from tests import utils as tutils
+
+IS_GIHTUB_ACTIONS = os.getenv("GITHUB_ACTIONS") == "true"
+
 
 def test_apt_all(client):
     """Test apt_all endpoint."""
@@ -25,6 +32,19 @@ def test_apt_new_post(client, appointment_finput):
     assert response.status_code == 302
 
 
+@pytest.mark.skipif(IS_GIHTUB_ACTIONS, reason="Test doesn't work in Github Actions.")
+def test_apt_new_post_email(app, appointment_finput):
+    """Test apt_new endpoint with email."""
+    app.config["EMAIL"] = "gonzalo.garcia@sjd.es"
+    client = app.test_client()
+    response = client.post("/participants/1/appointment_new", data=appointment_finput)
+    assert response.status_code == 302
+    time.sleep(20)
+    email = tutils.check_email_received()
+    assert email
+    assert "Appointment 1:" in email["subject"]
+
+
 def test_apt_mod(client, appointment_finput_mod):
     """Test apt_all endpoint."""
     ppt_id = appointment_finput_mod["inputId"]
@@ -42,3 +62,21 @@ def test_apt_mod_post(client, appointment_finput_mod):
         data=appointment_finput_mod,
     )
     assert response.status_code == 302
+
+
+@pytest.mark.skipif(IS_GIHTUB_ACTIONS, reason="Test doesn't work in Github Actions.")
+def test_apt_mod_post_email(app, appointment_finput_mod):
+    """Test apt_post endpoint with email."""
+    app.config["EMAIL"] = "gonzalo.garcia@sjd.es"
+    client = app.test_client()
+    ppt_id = appointment_finput_mod["inputId"]
+    apt_id = appointment_finput_mod["inputAptId"]
+    response = client.post(
+        f"/participants/{ppt_id}/{apt_id}/appointment_modify",
+        data=appointment_finput_mod,
+    )
+    assert response.status_code == 302
+    time.sleep(20)
+    email = tutils.check_email_received()
+    assert email
+    assert "Appointment 1:" in email["subject"]
